+++ build.ps1
@@ -3,7 +3,6 @@
               .\src\custom.js,`
               .\src\feature-modules\chat.js,`
               .\src\feature-modules\pm.js,`
-              .\src\feature-modules\rng.js,`
               .\src\feature-modules\modding.js,`
               .\src\core-modules\settings.js,`
               .\src\core-modules\about.js,`