+++ build.sh
@@ -5,7 +5,6 @@ cat ./src/header.js \
     ./src/custom.js \
     ./src/feature-modules/chat.js \
     ./src/feature-modules/session.js \
-    ./src/feature-modules/rng.js \
     ./src/feature-modules/blocking.js \
     ./src/core-modules/settings.js \
     ./src/core-modules/about.js \